+++ app/page.tsx
@@ -1,55 +1,56 @@
 import React from 'react';
-import Chat from "@/components/Chat";
-
-import { createClient } from '@/utils/supabase/server'
-import { redirect } from 'next/navigation'
-import { Channel } from '@/utils/types/types'
-import NavBar from "@/components/NavBar";
-import {Sidebar} from "@/components/Catalyst/sidebar";
+import { createClient } from '@/utils/supabase/server';
+import { redirect } from 'next/navigation';
+import NavBar  from "@/components/NavBar";
 import {StackedLayout} from "@/components/Catalyst/stacked-layout";
+import { Channel } from '@/utils/types/types'
+import {Sidebar} from "@/components/Catalyst/sidebar";
 import {
-  fetchAllChannelsForCurrentUser, 
-  fetchAllPreviousMessages,
-  fetchChannelUsers 
+    fetchAllChannelsForCurrentUser,
+    fetchAllPreviousMessages,
+    fetchChannelUsers
 } from '@/utils/api/api';
-
+import Master from "@/components/Master";
+// Ensure AdminPanelComponent is a proper component
 
 /**
- * The entry point to our application.
+ * The entry point to our Admin Panel.
  * Handles authentication and data fetching.
- * @returns Chat component for real-time messaging
+ * @returns AdminPanel component
  */
-export default async function Page() {
-  // Check if user is authenticated
-  const supabase = await createClient()
-  const {
-      data: { user },
-  } = await supabase.auth.getUser()
+export default async function AdminPanelPage() {
+    // Initialize Supabase
+    const supabase = await createClient();
 
-  if (!user) {
-      return redirect('/login')
-  }
+    // Check if user is authenticated
+    const { data: { user } } = await supabase.auth.getUser();
 
-  // fetch all channels the user is subscribed to
-  const channels: Channel[] = await fetchAllChannelsForCurrentUser(user)
+    if (!user) {
+        return redirect('/login');
+    }
 
-  // fetch previous messages
-  const previousMessages = await fetchAllPreviousMessages(channels)
+    // CHANNEL BAR HELP
+    // fetch all channels the user is subscribed to
+    // fetch all channels the user is subscribed to
+    const channels: Channel[] = await fetchAllChannelsForCurrentUser(user)
 
-  // fetch all users for all channels a user is subscribed to
-  const channelUsers = await fetchChannelUsers(channels)
+    // fetch previous messages
+    const previousMessages = await fetchAllPreviousMessages(channels)
 
-  return (
-      <StackedLayout
-          navbar={<NavBar/>}
-          sidebar={<Sidebar />}
-          >
-        <Chat
-            user={user}
-            previousMessages={previousMessages}
-            channels={channels}
-            channelUsers={channelUsers}
-        />
-      </StackedLayout>
-  )
+    // fetch all users for all channels a user is subscribed to
+    const channelUsers = await fetchChannelUsers(channels)
+
+    return (
+        <StackedLayout
+            navbar={<NavBar />}
+            sidebar={<Sidebar />}
+        >
+            <Master
+                user={user}
+                previousMessages={previousMessages}
+                channels={channels}
+                channelUsers={channelUsers}
+            />
+        </StackedLayout>
+    );
 }
